requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dataclasses==0.6
Flask==1.0.3
Flask-Cors==3.0.8
google-streetview==1.2.9
gunicorn==19.9.0
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
kwconfig==1.1.7
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
prometheus-client==0.7.0
pylint==2.3.1
python-dateutil==2.8.0
python-googlegeocoder==0.4.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rope==0.14.0
six==1.12.0
toml==0.10.0
torch==1.2.0
torchvision==0.4.0
typed-ast==1.3.5
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.1
